4.22

+++ controllers/blogs.js
@@ -1,8 +1,5 @@
 const blogsRouter = require("express").Router()
-const jwt = require("jsonwebtoken")
 const Blog = require("../models/blog")
-const User = require("../models/user")
-
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate("user", { username: 1, name: 1, id: 1 })
@@ -10,9 +7,10 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-
-    const user = await User.findById(decodedToken.id)
+    const user = request.user
+    if (!user) {
+        return response.status(401).json({error: "no user token given"})
+    }
     const blog = new Blog({ ...request.body, user: user._id })
 
     const result = await blog.save()
@@ -23,20 +21,19 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-
-    const user = await User.findById(decodedToken.id)
-    console.log(request.params.id)
+    const user = request.user
+    if (!user) {
+        return response.status(401).json({error: "no user token given"})
+    }
     const blog = await Blog.findById(request.params.id)
-    console.log(blog)
-    if (!blog){
+    if (!blog) {
         return response.status(204).end()
     }
-    if (blog.user._id.toString() === user._id.toString()){
-        await Blog.deleteOne({_id: request.params.id})
+    else if (blog.user._id.toString() === user._id.toString()) {
+        await Blog.deleteOne({ _id: request.params.id })
         return response.status(204).end()
-    }
-    response.status(401).json({error: "user not authorized to delete this blog"})
+    } else
+        response.status(401).json({ error: "user not authorized to delete this blog" })
 })
 
 blogsRouter.put("/:id", async (request, response) => {
